requirements.txt
absl-py==0.9.0
astroid==2.4.0
attrs==19.3.0
backcall==0.1.0
bleach==3.1.4
cachetools==4.1.0
certifi==2020.4.5.1
chardet==3.0.4
cloudpickle==1.4.1
cycler==0.10.0
Cython==0.29.17
decorator==4.4.2
defusedxml==0.6.0
detectron2==0.1.2+cu101
entrypoints==0.3
environs==7.4.0
feather-format==0.4.1
flake8==3.7.9
future==0.18.2
fvcore==0.1.dev200506
google-auth==1.14.2
google-auth-oauthlib==0.4.1
grpcio==1.28.1
idna==2.9
importlib-metadata==1.5.0
ipykernel==5.1.4
ipython==7.13.0
ipython-genutils==0.2.0
isort==4.3.21
jedi==0.17.0
Jinja2==2.11.2
joblib==0.14.1
jsonschema==3.2.0
jupyter-client==6.1.3
jupyter-core==4.6.3
kaggle==1.5.6
kiwisolver==1.2.0
lazy-object-proxy==1.4.3
Markdown==3.2.2
MarkupSafe==1.1.1
marshmallow==3.6.0
matplotlib==3.1.3
mccabe==0.6.1
mistune==0.8.4
mkl-fft==1.0.15
mkl-random==1.1.0
mkl-service==2.3.0
mock==4.0.2
nbconvert==5.6.1
nbformat==5.0.6
notebook==6.0.3
numpy==1.18.1
oauthlib==3.1.0
olefile==0.46
opencv-python==4.2.0.34
pandas==1.0.3
pandocfilters==1.4.2
parso==0.7.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.1.2
portalocker==1.7.0
prometheus-client==0.7.1
prompt-toolkit==3.0.4
protobuf==3.11.3
ptyprocess==0.6.0
pyarrow==0.17.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycocotools==2.0
pycodestyle==2.5.0
pydot==1.4.1
pyflakes==2.1.1
Pygments==2.6.1
pylint==2.5.0
pyparsing==2.4.7
pyrsistent==0.16.0
python-dateutil==2.8.1
python-dotenv==0.13.0
python-slugify==4.0.0
pytz==2020.1
PyYAML==5.3.1
pyzmq==18.1.1
requests==2.23.0
requests-oauthlib==1.3.0
rsa==4.0
scikit-learn==0.23.0
scipy==1.4.1
seaborn==0.10.1
Send2Trash==1.5.0
sip==4.19.13
six==1.14.0
tabulate==0.8.7
tensorboard==2.2.1
tensorboard-plugin-wit==1.6.0.post3
termcolor==1.1.0
terminado==0.8.3
testpath==0.4.4
text-unidecode==1.3
threadpoolctl==2.0.0
toml==0.10.0
torch==1.5.0+cu101
torchvision==0.6.0+cu101
tornado==6.0.4
tqdm==4.46.0
traitlets==4.3.3
urllib3==1.24.3
wcwidth==0.1.9
webencodings==0.5.1
Werkzeug==1.0.1
wrapt==1.11.2
yacs==0.1.7
zipp==3.1.0
